requirementes.txt
asgiref==3.6.0
Django==4.1.6
django-cors-headers
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
itsdangerous
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
